all errors fixed

+++ src/main/java/org/ncl/cloudcomputing/ttp/TTP.java
@@ -31,17 +31,22 @@ public class TTP extends AWSBase implements Runnable {
     	messageAttributes.put("sig-alice", new MessageAttributeValue().withDataType("Binary").withBinaryValue(ByteBuffer.wrap(sigAlice)));
     	messageAttributes.put("transaction-id", new MessageAttributeValue().withDataType("String").withStringValue(transactionId));
     	
+    	messageAttributes.put("message-status", new MessageAttributeValue().withDataType("Number").withStringValue(MessageStatus.TTP_to_Bob.getValue().toString()));
+    	
     	SendMessageRequest request = new SendMessageRequest();
 	    request.withMessageAttributes(messageAttributes);
 	    request.setMessageBody("1");
 	    this.amazonBobQueue.sendMessage(request, MessageStatus.TTP_to_Bob);
 	}
 	
-	private void sendDocumentKeyToBob(String transactionId, String docKey) {
+	private void sendDocumentKeyToBob(String transactionId, String docKey, String filename) {
 		Map<String, MessageAttributeValue> messageAttributes = new HashMap<String, MessageAttributeValue>();
 
     	messageAttributes.put("transaction-id", new MessageAttributeValue().withDataType("String").withStringValue(transactionId));
     	messageAttributes.put("doc-key", new MessageAttributeValue().withDataType("String").withStringValue(docKey));
+    	messageAttributes.put("file-name", new MessageAttributeValue().withDataType("String").withStringValue(filename));
+    	
+    	messageAttributes.put("message-status", new MessageAttributeValue().withDataType("Number").withStringValue(MessageStatus.TTP_to_Bob_doc.getValue().toString()));
     	
     	SendMessageRequest request = new SendMessageRequest();
 	    request.withMessageAttributes(messageAttributes);
@@ -54,6 +59,8 @@ public class TTP extends AWSBase implements Runnable {
 
     	messageAttributes.put("transaction-id", new MessageAttributeValue().withDataType("String").withStringValue(transactionId));
     	messageAttributes.put("sig-bob", new MessageAttributeValue().withDataType("Binary").withBinaryValue(ByteBuffer.wrap(sigBob)));
+    	
+    	messageAttributes.put("message-status", new MessageAttributeValue().withDataType("Number").withStringValue(MessageStatus.TTP_to_Alice.getValue().toString()));
     	
     	SendMessageRequest request = new SendMessageRequest();
 	    request.withMessageAttributes(messageAttributes);
@@ -70,41 +77,49 @@ public class TTP extends AWSBase implements Runnable {
 
 	public void run() {
 		while (true) {
+			System.out.println("Receiving messages...");
+			
 			List<Message> messages = this.amazonTTPQueue.receiveMessages();
+			
+			System.out.println(messages.size() + " messages received.");
+			
 			for (Message message : messages) {
-				String strMessageStatus = message.getAttributes().get("message-status").toString();
+				String strMessageStatus = message.getMessageAttributes().get("message-status").getStringValue();
 				Integer messageStatus = Integer.parseInt(strMessageStatus);
 				
 				Logger.logReceiveMessageOnSucceed(messageStatus);
 				
 				if (messageStatus == MessageStatus.Alice_to_TTP.getValue()) {
-					String transactionId = message.getAttributes().get("transaction-id").toString();
+					String transactionId = message.getMessageAttributes().get("transaction-id").getStringValue();
+					byte[] sigAlice = message.getMessageAttributes().get("sig-alice").getBinaryValue().array();
+					String docKey = message.getMessageAttributes().get("doc-key").getStringValue();
+					String filename = message.getMessageAttributes().get("file-name").getStringValue();
 					
-					byte[] sigAlice = message.getAttributes().get("sig-alice").getBytes();
-					String docKey = message.getAttributes().get("doc-key").toString();
+					TransactionItem item = new TransactionItem(transactionId, sigAlice, docKey, filename);
+					TransactionItem result = this.transactionRepo.insert(item);
 					
-					TransactionItem item = new TransactionItem(transactionId, sigAlice, docKey);
-					this.transactionRepo.insert(item);
-					
-					this.sendMessageToBob(transactionId, sigAlice);
+					if (result != null)
+						this.sendMessageToBob(transactionId, sigAlice);
 				}
 				else if (messageStatus == MessageStatus.Bob_to_TTP.getValue()) {
-					String transactionId = message.getAttributes().get("transaction-id").toString();
-					byte[] sigBob = message.getAttributes().get("sig-bob").getBytes();
+					String transactionId = message.getMessageAttributes().get("transaction-id").getStringValue();
+					byte[] sigBob = message.getMessageAttributes().get("sig-bob").getBinaryValue().array();
 					
 					// WE NEED TO KNOW IF SIGBOB SENT BY BOB IS CORRECT
 					/* to do this we need Bob's PublicKey.. could send it in the message or get Bob to transmit it to TTP another way? */
 					
 					TransactionItem transaction = this.transactionRepo.getItemById(transactionId);
 					
-					this.sendDocumentKeyToBob(transactionId, transaction.getDocumentKey());
-					this.sendMessageToAlice(transactionId, sigBob);
-					
-					this.transactionRepo.delete(transactionId);
+					if (transaction != null) {
+						this.sendDocumentKeyToBob(transactionId, transaction.getDocumentKey(), transaction.getFilename());
+						this.sendMessageToAlice(transactionId, sigBob);
+						this.transactionRepo.delete(transactionId);
+					}
 				}
 			}
 			
-			this.amazonTTPQueue.deleteMessages();
+			//if (messages.size() != 0)
+				//this.amazonTTPQueue.deleteMessages();
 			
 			try {
 				Thread.sleep(1000);
